refactor http error handling

+++ apps/backend/src/modules/bookModule/application/commandHandlers/deleteAuthorCommandHandler/deleteAuthorCommandHandlerImpl.ts
@@ -2,6 +2,7 @@ import {
   type DeleteAuthorCommandHandler,
   type DeleteAuthorCommandHandlerPayload,
 } from './deleteAuthorCommandHandler.js';
+import { ResourceNotFoundError } from '../../../../../common/errors/resourceNotFoundError.js';
 import { type LoggerService } from '../../../../../libs/logger/services/loggerService/loggerService.js';
 import { type AuthorRepository } from '../../../domain/repositories/authorRepository/authorRepository.js';
 
@@ -18,6 +19,15 @@ export class DeleteAuthorCommandHandlerImpl implements DeleteAuthorCommandHandle
       message: 'Deleting author...',
       authorId,
     });
+
+    const existingAuthor = await this.authorRepository.findAuthor({ id: authorId });
+
+    if (!existingAuthor) {
+      throw new ResourceNotFoundError({
+        resource: 'Author',
+        id: authorId,
+      });
+    }
 
     await this.authorRepository.deleteAuthor({ id: authorId });
