+++ apps/backend/src/modules/userModule/application/commandHandlers/deleteUserCommandHandler/deleteUserCommandHandlerImpl.ts
@@ -1,4 +1,5 @@
 import { type DeleteUserCommandHandler, type DeleteUserCommandHandlerPayload } from './deleteUserCommandHandler.js';
+import { ResourceNotFoundError } from '../../../../../common/errors/resourceNotFoundError.js';
 import { type LoggerService } from '../../../../../libs/logger/services/loggerService/loggerService.js';
 import { type UserRepository } from '../../../domain/repositories/userRepository/userRepository.js';
 
@@ -15,6 +16,15 @@ export class DeleteUserCommandHandlerImpl implements DeleteUserCommandHandler {
       message: 'Deleting User...',
       userId,
     });
+
+    const existingUser = await this.userRepository.findUser({ id: userId });
+
+    if (!existingUser) {
+      throw new ResourceNotFoundError({
+        resource: 'User',
+        userId,
+      });
+    }
 
     await this.userRepository.deleteUser({ id: userId });
